backend-server/requirements.txt
aiobotocore==2.4.2
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
altair==4.2.2
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.0
bleach==6.0.0
blessed==1.20.0
blinker==1.5
bokeh==2.4.3
botocore==1.27.59
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
croniter==1.3.8
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.1
django-bootstrap-v5==1.0.11
django-cors-headers==4.2.0
django-filter==23.2
django-jquery==3.1.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
dnspython==2.3.0
docker==6.0.1
docstring-parser==0.15
email-validator==1.3.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.100.0
fastjsonschema==2.16.3
filelock==3.10.0
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.31
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
hydra-core==1.3.2
idna==3.4
imageio==2.27.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
inquirer==3.1.3
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.0
jmespath==1.0.1
jsonargparse==4.20.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lazy_loader==0.2
lightning==2.0.5
lightning-api-access==0.0.5
lightning-cloud==0.5.37
lightning-fabric==2.0.0
lightning-utilities==0.8.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
omegaconf==2.3.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.0.10
ordered-set==4.1.0
orjson==3.8.8
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pyct==0.5.0
pydantic==1.10.6
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
readchar==4.0.5
redis==4.5.2
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.2
s3fs==2022.11.0
scikit-image==0.20.0
scipy==1.10.1
selenium==4.8.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
streamlit==1.20.0
sympy==1.11.1
tensorboardX==2.6
terminado==0.17.1
tifffile==2023.4.12
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typeshed-client==2.2.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
ujson==5.7.0
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.21.1
validators==0.20.0
watchdog==3.0.0
watchfiles==0.18.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
whitenoise==6.5.0
widgetsnbextension==4.0.5
wrapt==1.15.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
